Player basics
Created a PlayerController to move and rotate the ship.

Created a Shooter script to manage the player shooting

Created a ScreenWraper script so that some elements changes position when on the edge of the screen

Created a basic ObjectPooler script to manage pooling (for now only for bullets, will extend later on)

Imported DOTween for later use.